+++ Android/app/src/main/java/use_case/login/LoginOutputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.boundary.output;
+package use_case.login;
 
 import constant.uimessage.LoginResult;
